requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
dj-database-url==2.1.0
python-decouple==3.8
Pillow==10.2.0
gunicorn==21.2.0
openai>=1.14.0
